+++ frontend/src/main.tsx
@@ -1,8 +1,9 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+import { App } from "./App.tsx";
 import "./index.css";
-import App from "./App.tsx";
 
+// biome-ignore lint/style/noNonNullAssertion: vite initした時にこの書き方なので、良いでしょう
 createRoot(document.getElementById("root")!).render(
 	<StrictMode>
 		<App />
